refactor(api): supprimer le doublon du module mergeOperators refers to https://trello.com/c/cCbgradg/2675
